backend/requirements.txt
fastapi==0.135.1
python-dotenv==1.2.2
uvicorn==0.41.0
websockets==16.0
pydantic
google-genai